+++ crates/block-producer/src/server/mod.rs
@@ -75,8 +75,8 @@ pub struct BlockProducer {
 impl BlockProducer {
     /// Serves the block-producer RPC API, the batch-builder and the block-builder.
     ///
-    /// Note: Executes in place (i.e. not spawned) and will run indefinitely until
-    ///       a fatal error is encountered.
+    /// Executes in place (i.e. not spawned) and will run indefinitely until a fatal error is
+    /// encountered.
     #[allow(clippy::too_many_lines)]
     pub async fn serve(self) -> anyhow::Result<()> {
         info!(target: COMPONENT, endpoint=?self.block_producer_address, store=%self.store_url, "Initializing server");