+++ crates/proto/build.rs
@@ -10,6 +10,7 @@ use miden_node_proto_build::{
     store_ntx_builder_api_descriptor,
     store_rpc_api_descriptor,
     store_shared_api_descriptor,
+    validator_api_descriptor,
 };
 use miette::{Context, IntoDiagnostic};
 use tonic_build::FileDescriptorSet;
@@ -46,6 +47,7 @@ fn main() -> miette::Result<()> {
     generate_bindings(store_shared_api_descriptor(), &dst_dir)?;
     generate_bindings(block_producer_api_descriptor(), &dst_dir)?;
     generate_bindings(remote_prover_api_descriptor(), &dst_dir)?;
+    generate_bindings(validator_api_descriptor(), &dst_dir)?;
 
     generate_mod_rs(&dst_dir).into_diagnostic().wrap_err("generating mod.rs")?;