+++ crates/validator/src/lib.rs
@@ -1,0 +1,9 @@
+mod server;
+
+pub use server::Validator;
+
+// CONSTANTS
+// =================================================================================================
+
+/// The name of the validator component.
+pub const COMPONENT: &str = "miden-validator";